Product ID changed from int to Long

+++ SuperDiaClient/ejbModule/br/com/superdia/modelo/Item.java
@@ -14,7 +14,7 @@ public class Item implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     
     @ManyToOne
     private Produto produto;
@@ -24,7 +24,7 @@ public class Item implements Serializable {
 
     public Item() {}
 
-    public Item(int id, Produto produto, Integer quantidade, Double valorUnitario) {
+    public Item(Long id, Produto produto, Integer quantidade, Double valorUnitario) {
 		super();
 		this.id = id;
 		this.produto = produto;
@@ -32,11 +32,11 @@ public class Item implements Serializable {
 		this.valorUnitario = valorUnitario;
 	}
     
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
